requirements.txt
streamlit
openai
python-dotenv
fastapi
pyperclip
